requirements.txt
flask==2.3.3
flask-cors==4.0.0
docx2txt==0.8
python-docx==0.8.11
easyocr==1.7.0
pillow==10.0.0
numpy==1.24.3
